+++ speechflow-cli/src/lib.d.ts
@@ -9,3 +9,20 @@ declare module "node:stream" {
     export function compose (...streams: Stream[]): Duplex
 }
 
+/*  type definitions for AudioWorkletProcessor  */
+interface AudioWorkletProcessor {
+    readonly port: MessagePort
+    process(
+        inputs: Float32Array[][],
+        outputs: Float32Array[][],
+        parameters: Record<string, Float32Array>
+    ): boolean
+}
+declare const AudioWorkletProcessor: {
+    prototype: AudioWorkletProcessor
+    new(): AudioWorkletProcessor
+}
+declare function registerProcessor(
+    name: string,
+    processorCtor: new () => AudioWorkletProcessor
+): void